requirements.txt
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.3.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
python-telegram-bot==20.8
pytz==2024.2
requests==2.32.0
